+++ RELEASE_NOTES.md
@@ -1,4 +1,4 @@
 ### Fixes
 
- - Add `brand_name` property to Product (Issue #99, thanks @guillaume-sarramegna) 
+ - Fixes issue with Product Option Value API